fix(workspace): split node APIs from browser barrel
