+++ apps/opensidian/src/lib/opensidian/script.ts
@@ -1,14 +1,16 @@
 import { EPICENTER_API_URL } from '@epicenter/constants/apps';
 import {
 	attachSync,
-	attachYjsLogReader,
-	findEpicenterDir,
-	hashClientId,
 	type ProjectDir,
 	toWsUrl,
 	type WebSocketImpl,
-	yjsPath,
 } from '@epicenter/workspace';
+import {
+	attachYjsLogReader,
+	findEpicenterDir,
+	hashClientId,
+	yjsPath,
+} from '@epicenter/workspace/node';
 import { openOpensidian as openOpensidianDoc } from './index.js';
 
 export function openOpensidian({
