+++ packages/cli/src/util/common-options.ts
@@ -7,7 +7,7 @@
 
 import { resolve } from 'node:path';
 
-import { findEpicenterDir } from '@epicenter/workspace';
+import { findEpicenterDir } from '@epicenter/workspace/node';
 import type { Options } from 'yargs';
 
 export function resolveProjectDir(start: string): string {
